pyproject.toml
[project]
name = "snowflake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.4",
    "simple-salesforce>=1.12.6",
    "snowflake-connector-python>=3.16.0",
]
